requirements.txt
gradio>=4.0
charset-normalizer>=3.0
openai>=1.0
python-dotenv>=1.0
pyinstaller>=6.0
